requirements.txt
streamlit
pageindex
google-generativeai
requests
python-dotenv